feat: Add search value on list attribute set instances.
